added browser support for bitauth
