+++ package.json
@@ -5,12 +5,27 @@
     "name": "Patrick Nagurny",
     "email": "patrick@bitpay.com"
   },
-  "version": "0.1.0",
+  "contributors": [
+    {
+      "name": "Gordon Hall",
+      "email": "gordon@bitpay.com"
+    }
+  ],
+  "scripts": {
+    "make-dist": "sh scripts/make-dist.sh",
+    "postinstall": "npm run make-dist"
+  },
+  "main": "index.js",
+  "version": "0.1.1",
   "dependencies": {
     "bitcore": ">= 0.1.9",
     "request": "^2.36.0",
     "express": "^4.3.1",
     "base58-native": "^0.1.4",
     "body-parser": "^1.2.0"
+  },
+  "devDependencies": {
+    "uglify-js": "~2.4.14",
+    "browserify": "~4.1.11"
   }
 }
